Move confirmation methods into ConfirmationsController
